requirements.txt
asgiref==3.7.2
attrs==23.1.0
boto3==1.28.82
botocore==1.31.82
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==4.3.0
django-environ==0.11.2
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
jmespath==1.0.1
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.7.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
